% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBHC.R
\name{mbhc.find.cut.off}
\alias{mbhc.find.cut.off}
\title{find cut off}
\usage{
mbhc.find.cut.off(mbhc.output)
}
\arguments{
\item{mbhc.output}{The output of mbhc.single is the input of find.cut.off}
}
\description{
This function finds the optimal cluster assignment for the data.
}
