% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBHC.R
\name{mbhc.multiple.cut}
\alias{mbhc.multiple.cut}
\title{MBHC multiple}
\usage{
mbhc.multiple.cut(arg.data, x.el.cut = data.frame(), sampleNum = 1)
}
\arguments{
\item{arg.data, x.el.cut, sampleNum}{Arg.data should be a dataframe with IDs in the last column. Each column correspond to a sample. X.el.cut is the output structure of mbhc.reduce which performs a prliminary clustering. it can be empty. SampleNum is the column of the primary sample that clustering needs to be run on.}
}
\description{
This function performs the clustering on multiple samples.
}
