% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBHC.R
\name{mbhc.prepdata}
\alias{mbhc.prepdata}
\title{MBHC preparing the data}
\usage{
mbhc.prepdata(arg.data)
}
\arguments{
\item{arg.data}{the input data. In the dataframe format.}
}
\description{
This function creates ID for the data and preprocess the data for the MBHC.
}
