% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBHC.R
\name{mbhc.single}
\alias{mbhc.single}
\title{Model based hc}
\usage{
mbhc.single(arg.data, sampleNum = 1)
}
\arguments{
\item{arg.data, sampleNum}{Arg.data is the output of mbhc.prepdata function with IDs added. sampleNum for single sample should be 1}
}
\description{
This function performs the model-based hierarchical clustering
}
